#52 入力フォームを子コンポーネントで分割
